+++ R/ggpredict.R
@@ -128,24 +128,30 @@
 #'   exact values, for instance `condition = c(covariate1 = 20, covariate2 = 5)`.
 #'   See 'Examples'.
 #' @param interval Type of interval calculation, can either be `"confidence"`
-#'   (default) or `"prediction"`. May be abbreviated. Unlike
-#'   *confidence intervals*, *prediction intervals* include the
-#'   residual variance (sigma^2). This argument is ignored for mixed models,
-#'   as `interval = "prediction"` is equivalent to `type = "random"`
-#'   (and `interval = "confidence"` is equivalent to `type = "fixed"`).
-#'   Note that prediction intervals are not available for all models, but only
-#'   for models that work with [`insight::get_sigma()`].
-#' @param vcov.fun String, indicating the name of the `vcov*()`-function
-#'    from the **sandwich** or **clubSandwich**-package, e.g.
-#'    `vcov.fun = "vcovCL"`, which is used to compute (cluster) robust
-#'    standard errors for predictions. If `NULL`, standard errors (and
-#'    confidence intervals) for predictions are based on the standard errors as
-#'    returned by the `predict()`-function. **Note** that probably not
-#'    all model objects that work with `ggpredict()` are also supported
-#'    by the **sandwich** or **clubSandwich**-package.
+#'   (default) or `"prediction"`. May be abbreviated. Unlike *confidence intervals*,
+#'   *prediction intervals* include the residual variance (sigma^2). For mixed
+#'   models, `interval = "prediction"` is the default for `type = "random"`.
+#'   When `type = "fixed"`, the default is `interval = "confidence"`. Note that
+#'   prediction intervals are not available for all models, but only for models
+#'   that work with [`insight::get_sigma()`].
+#' @param vcov.fun Variance-covariance matrix used to compute uncertainty
+#'   estimates (e.g., for confidence intervals based on robust standard errors).
+#'   This argument accepts a covariance matrix, a function which returns a
+#'   covariance matrix, or a string which identifies the function to be used to
+#'   compute the covariance matrix.
+#'   * A covariance matrix
+#'   * A function which returns a covariance matrix (e.g., `stats::vcov()`)
+#'   * A string which indicates the name of the `vcov*()`-function from the
+#'     **sandwich** or **clubSandwich**-package, e.g. `vcov.fun = "vcovCL"`,
+#'     which is used to compute (cluster) robust standard errors for predictions.
+#'     If `NULL`, standard errors (and confidence intervals) for predictions are
+#'     based on the standard errors as returned by the `predict()`-function.
+#'     **Note** that probably not all model objects that work with `ggpredict()`
+#'     are also supported by the **sandwich** or **clubSandwich**-package.
 #' @param vcov.type Character vector, specifying the estimation type for the
-#'    robust covariance matrix estimation (see `?sandwich::vcovHC`
-#'    or `?clubSandwich::vcovCR` for details).
+#'   robust covariance matrix estimation (see `?sandwich::vcovHC`
+#'   or `?clubSandwich::vcovCR` for details). Only used when `vcov.fun` is a
+#'   character string.
 #' @param vcov.args List of named vectors, used as additional arguments that
 #'    are passed down to `vcov.fun`.
 #' @param verbose Toggle messages or warnings.
@@ -461,7 +467,7 @@
 ggpredict <- function(model,
                       terms,
                       ci.lvl = 0.95,
-                      type = "fe",
+                      type = "fixed",
                       typical = "mean",
                       condition = NULL,
                       back.transform = TRUE,
@@ -469,7 +475,7 @@ ggpredict <- function(model,
                       vcov.fun = NULL,
                       vcov.type = NULL,
                       vcov.args = NULL,
-                      interval = "confidence",
+                      interval,
                       verbose = TRUE,
                       ...) {
   # check arguments
@@ -479,8 +485,6 @@ ggpredict <- function(model,
                                       "sim", "simulate", "surv", "survival", "cumhaz",
                                       "cumulative_hazard", "sim_re", "simulate_random",
                                       "debug"))
-  interval <- match.arg(interval, choices = c("confidence", "prediction"))
-  model.name <- deparse(substitute(model))
 
   type <- switch(
     type,
@@ -498,6 +502,16 @@ ggpredict <- function(model,
     "simulate_random" = "sim_re",
     type
   )
+
+  if (missing(interval)) {
+    if (type %in% c("re", "re.zi")) {
+      interval <- "prediction"
+    } else {
+      interval <- "confidence"
+    }
+  }
+  interval <- match.arg(interval, choices = c("confidence", "prediction"))
+  model.name <- deparse(substitute(model))
 
   # check if terms are a formula
   if (!missing(terms) && !is.null(terms) && inherits(terms, "formula")) {
@@ -710,6 +724,7 @@ ggpredict_helper <- function(model,
     ci.lvl = ci.lvl,
     untransformed.predictions = untransformed.predictions,
     back.transform = back.transform,
-    response.transform = response.transform
+    response.transform = response.transform,
+    verbose = verbose
   )
 }