+++ R/post_processing_labels.R
@@ -12,7 +12,8 @@
                                     ci.lvl = 0.95,
                                     untransformed.predictions = NULL,
                                     back.transform = FALSE,
-                                    response.transform = NULL) {
+                                    response.transform = NULL,
+                                    verbose = TRUE) {
   # get axis titles and labels
   all.labels <- .get_axis_titles_and_labels(
     model,
@@ -21,7 +22,8 @@
     fun = .get_model_function(model),
     model_info = model_info,
     no.transform = FALSE,
-    type = type
+    type = type,
+    at_list = at_list
   )
 
   # set attributes with necessary information
@@ -46,6 +48,7 @@
     type = type,
     untransformed.predictions = untransformed.predictions,
     back.transform = back.transform,
-    response.transform = response.transform
+    response.transform = response.transform,
+    verbose = verbose
   )
 }